income_expense_API/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
loguru==0.7.2
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
